base_code/engine.py
import socket
import threading
import json
import logging
from typing import Dict, Any
import yaml
import queue
import multiprocessing

class DistributedProcessingEngine:
    def __init__(self, config_path: str = 'config.yaml'):
        """Initialize distributed processing engine
        Args:
            config_path (str): Path to configuration file"""
        # Load configuration
        with open(config_path, 'r') as config_file:
            self.config = yaml.safe_load(config_file)
        
        # Setup logging
        self._configure_logging()
        
        # Task queue for distributed processing
        self.task_queue = multiprocessing.Queue()
        self.result_queue = multiprocessing.Queue()
        
        # Network configuration
        self.host = self.config['network']['host']
        self.port = self.config['network']['port']

    def _configure_logging(self):
        """Configure logging for the processing engine"""
        logging_config = self.config.get('logging', {})
        log_level = getattr(logging, logging_config.get('level', 'INFO').upper())
        
        logging.basicConfig(
            level=log_level,
            format=logging_config.get('format', '%(asctime)s - %(levelname)s - %(message)s')
        )
        self.logger = logging.getLogger(__name__)

    def start_server(self):
        """Start socket server for distributed task management"""
        server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
        server_socket.bind((self.host, self.port))
        server_socket.listen(self.config['network']['max_connections'])
        
        self.logger.info(f"Server listening on {self.host}:{self.port}")
        
        while True:
            client_socket, address = server_socket.accept()
            self.logger.info(f"Connection from {address}")
            
            # Handle each client in a separate thread
            client_thread = threading.Thread(
                target=self._handle_client, 
                args=(client_socket,)
            )
            client_thread.start()

    def _handle_client(self, client_socket: socket.socket):
        """Handle individual client connections
        Args:
            client_socket (socket.socket): Connected client socket"""
        try:
            # Receive task data
            data = client_socket.recv(4096).decode('utf-8')
            task = json.loads(data)
            
            # Process task
            result = self._process_task(task)
            
            # Send result back to client
            response = json.dumps(result).encode('utf-8')
            client_socket.send(response)
        
        except Exception as e:
            self.logger.error(f"Client handling error: {e}")
        
        finally:
            client_socket.close()

    def _process_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Process a distributed task
        Args:
            task (Dict): Task details
        Returns:
            Dict: Processing result"""
        try:
            # Placeholder task processing logic
            self.task_queue.put(task)
            processed_result = {
                'status': 'success',
                'task_id': task.get('task_id'),
                'processed_at': str(datetime.now())
            }
            return processed_result
        
        except Exception as e:
            self.logger.error(f"Task processing error: {e}")
            return {
                'status': 'failed',
                'error': str(e)
            }

def main():
    engine = DistributedProcessingEngine()
    engine.start_server()

if __name__ == '__main__':
    main()