base_code/requirements.txt
# Core Python Libraries
python-dateutil==2.8.2
pyyaml==6.0.1
psutil==5.9.5

# Multiprocessing and Concurrency
concurrent-futures==3.4.0
multiprocessing-logging==0.3.1

# Network and Socket Programming
socketserver==0.4

# Optional: Advanced logging and monitoring
rich==10.16.2
prometheus-client==0.16.0

# Testing
pytest==7.3.1
pytest-cov==4.0.0